app/jurisbot/opensearch-ingestion/service-search/requirements.txt
boto3==1.34.0
pymysql==1.1.0
cryptography==41.0.7
sqlalchemy==2.0.23
opensearch-py==2.4.2
requests-aws4auth==1.2.3
pyyaml==6.0.1
python-dotenv==1.0.0
pandas==2.1.4
requests==2.31.0
urllib3==2.1.0